requirements.txt
PyGithub==2.6.1
requests==2.32.3